100GPT-server/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pillow
pytesseract
python-multipart
google-generativeai